pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reason2return"
version = "0.1.0"
description = "Structured reasoning for financial decision support using LLMs"
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = {text = "MIT"}
authors = [
    {name = "Justin Manning", email = "your.email@example.com"}
]
keywords = ["finance", "llm", "reinforcement-learning", "trading"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "numpy>=1.24,<2.0",
    "pandas>=2.0,<3.0",
    "torch>=2.0,<3.0",
    "scikit-learn>=1.3,<2.0",
    "jsonschema>=4.19,<5.0",
    "pydantic>=2.0,<3.0",
    "PyYAML>=6.0,<7.0",
    "python-json-logger>=2.0,<3.0",
    "python-dateutil>=2.8,<3.0",
    "tqdm>=4.66,<5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<8.0",
    "pytest-cov>=4.1,<5.0",
    "pytest-mock>=3.12,<4.0",
    "black>=23.10,<24.0",
    "ruff>=0.1,<1.0",
    "mypy>=1.6,<2.0",
    "jupyter>=1.0,<2.0",
    "ipykernel>=6.26,<7.0",
    "matplotlib>=3.8,<4.0",
]
api = [
    "fastapi>=0.104,<1.0",
    "uvicorn[standard]>=0.24,<1.0",
]
ml = [
    "mlflow>=2.8,<3.0",
]

[project.urls]
Homepage = "https://github.com/Sinnema1/reason2return"
Repository = "https://github.com/Sinnema1/reason2return"
Issues = "https://github.com/Sinnema1/reason2return/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["r2r*"]
exclude = ["tests*", "notebooks*", "configs*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.eggs
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["r2r"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "W",    # pycodestyle warnings
    "D",    # pydocstyle (docstrings)
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "S",    # flake8-bandit (security)
    "C90",  # mccabe (complexity)
    "NPY",  # numpy-specific rules
]
ignore = [
    "E501",  # Line too long (black handles this)
    "D100",  # Missing docstring in public module
    "D104",  # Missing docstring in public package
    "D203",  # 1 blank line required before class docstring (conflicts with D211)
    "D213",  # Multi-line docstring summary should start at the second line
    "S101",  # Use of assert detected (needed for tests)
    "S311",  # Standard pseudo-random generators are not suitable for cryptographic purposes
    "UP007", # Use X | Y for type annotations (Python 3.10+ syntax, TODO: enable in P1)
    "N806",  # Variable in function should be lowercase (R1/R2 are intentional reward names)
    "C901",  # Function is too complex (TODO: refactor complex functions in P1)
    "NPY002", # Legacy numpy random (TODO: migrate to Generator in P1)
]
per-file-ignores = { "tests/*" = ["D", "S"], "__init__.py" = ["F401"] }

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.mypy]
python_version = "3.10"
warn_return_any = false  # TODO P1: Enable after adding proper type hints
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_reexport = true
strict_equality = true
disallow_untyped_defs = false  # Start lenient, will tighten in P1
disallow_any_generics = false  # Start lenient, will tighten in P1
ignore_missing_imports = true

# Per-module overrides for gradual adoption
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
warn_return_any = false

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=r2r",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=20",  # Current: 22%, Target P0: 40%, P1: 70%, P2: 90%
    "--strict-markers",
]

[tool.bandit]
targets = ["r2r"]
exclude_dirs = ["tests"]
skips = ["B101"]  # Skip assert warnings in tests
